requirements.txt
mysql-connector-python
pymongo
tabulate